Move contents of preview.py to validate.py
